fix: update overview chart to display sum of averages, not average of all site data.
