+++ src/services/search.js
@@ -1,5 +1,8 @@
 export const DATE_HISTO = 'date_histogram#2';
 export const AVG = 'avg#1';
+
+export const MAX = 'max#1';
+
 export const TOTAL_REAL_POWER = 'Total Real Power';
 
 export const HOUR = 3600000;
@@ -76,8 +79,41 @@ export function parseSearchReturn(data) {
   });
 }
 
+export function parseMaxSearchReturn(data) {
+  return data.aggregations[DATE_HISTO].buckets.map((d) => {
+    return { date: new Date(Number(d.key)), values: d[MAX].value };
+  });
+}
+
 export function getAggregationValue(data, label) {
   return data !== undefined
     ? Math.round(data.aggregations[label].value * 10) / 10
     : 0;
 }
+
+export const parseStackedTimeSeriesData = function (data) {
+  const formattedData = [];
+  data.aggregations[DATE_HISTO].buckets.forEach((d) => {
+    const date = new Date(Number(d.key)).toISOString();
+    d['sterms#terms'].buckets.forEach((v) => {
+      formattedData.push({
+        date: date,
+        name: v.key,
+        avg: v[AVG] ? v[AVG].value : v['1'].value,
+      });
+    });
+  });
+  return formattedData;
+};
+
+export const parseAndCondenseStackedTimeSeriesData = function (data) {
+  const formattedData = [];
+  data.aggregations[DATE_HISTO].buckets.forEach((d) => {
+    const point = { date: new Date(Number(d.key)), values: 0 };
+    formattedData.push(point);
+    d['sterms#terms'].buckets.forEach((v) => {
+      point.values = point.values + (v[AVG] ? v[AVG].value : v['1'].value);
+    });
+  });
+  return formattedData;
+};
